requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy>=2.0.37
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
structlog==24.4.0
stripe==10.12.0
python-dotenv==1.0.1
anthropic>=0.40.0
